Refactors solution for Data project to depend on the core application instead of vice versa.
