+++ CoraCorpCM.App/CoraCorpCM.App.csproj
@@ -1,0 +1,11 @@
+<Project Sdk="Microsoft.NET.Sdk">
+
+  <PropertyGroup>
+    <TargetFramework>netcoreapp2.0</TargetFramework>
+  </PropertyGroup>
+
+  <ItemGroup>
+    <ProjectReference Include="..\CoraCorpCM.Domain\CoraCorpCM.Domain.csproj" />
+  </ItemGroup>
+
+</Project>
