+++ CoraCorpCM.Data/CoraCorpCM.Data.csproj
@@ -28,6 +28,7 @@
   </ItemGroup>
 
   <ItemGroup>
+    <ProjectReference Include="..\CoraCorpCM.App\CoraCorpCM.App.csproj" />
     <ProjectReference Include="..\CoraCorpCM.Domain\CoraCorpCM.Domain.csproj" />
   </ItemGroup>
