+++ CoraCorpCM.Web/Controllers/AccountController.cs
@@ -6,6 +6,7 @@ using Microsoft.AspNetCore.Authorization;
 using Microsoft.AspNetCore.Identity;
 using Microsoft.AspNetCore.Mvc;
 using Microsoft.Extensions.Logging;
+using CoraCorpCM.App.Interfaces;
 using CoraCorpCM.Domain;
 using CoraCorpCM.Domain.Models;
 using CoraCorpCM.Web.Services;
