+++ CoraCorpCM.Web/Utilities/ModelMapper.cs
@@ -1,4 +1,4 @@
-﻿using CoraCorpCM.Data;
+﻿using CoraCorpCM.App.Interfaces;
 using CoraCorpCM.Domain.Models;
 using CoraCorpCM.Web.ViewModels.CollectionViewModels;
 using System;
